refactor: move functions to more logical place
