+++ src/container.c
@@ -17,10 +17,9 @@
 #include "tagset.h"
 #include "ipc-server.h"
 #include "layer_shell.h"
+#include "workspace.h"
 
 static void add_container_to_workspace(struct container *con, struct workspace *ws);
-
-static void list_set_add_container_to_focus_stack(struct list_set *list_set, struct container *con);
 
 struct container *create_container(struct client *c, struct monitor *m, bool has_border)
 {
@@ -610,91 +609,62 @@ void resize_container(struct container *con, struct wlr_cursor *cursor, int offs
     container_damage(con, true);
 }
 
-void workspace_add_container_to_containers(struct workspace *ws, struct container *con, int i)
-{
-    assert(con != NULL);
-
-    DO_ACTION(
-        if (con->floating) {
-            g_ptr_array_insert(list_set->floating_containers, i, con);
-            continue;
-        }
-        if (con->hidden) {
-            g_ptr_array_insert(list_set->hidden_containers, i, con);
-            continue;
-        }
-        g_ptr_array_insert(list_set->tiled_containers, i, con);
-    );
-}
-
-static void list_set_add_container_to_focus_stack(struct list_set *list_set, struct container *con)
-{
-    if (con->client->type == LAYER_SHELL) {
-        switch (con->client->surface.layer->current.layer) {
-            case ZWLR_LAYER_SHELL_V1_LAYER_BACKGROUND:
-                g_ptr_array_insert(list_set->focus_stack_layer_background, 0, con);
-                break;
-            case ZWLR_LAYER_SHELL_V1_LAYER_BOTTOM:
-                g_ptr_array_insert(list_set->focus_stack_layer_bottom, 0, con);
-                break;
-            case ZWLR_LAYER_SHELL_V1_LAYER_TOP:
-                g_ptr_array_insert(list_set->focus_stack_layer_top, 0, con);
-                break;
-            case ZWLR_LAYER_SHELL_V1_LAYER_OVERLAY:
-                g_ptr_array_insert(list_set->focus_stack_layer_overlay, 0, con);
-                break;
-        }
-        return;
-    }
-    if (con->on_top) {
-        g_ptr_array_insert(list_set->focus_stack_on_top, 0, con);
-        return;
-    }
-    if (!con->focusable) {
-        g_ptr_array_insert(list_set->focus_stack_not_focusable, 0, con);
-        return;
-    }
-
-    g_ptr_array_insert(list_set->focus_stack_normal, 0, con);
-}
-
-void workspace_add_container_to_focus_stack(struct workspace *ws, struct container *con)
-{
-    DO_ACTION(
-        list_set_add_container_to_focus_stack(list_set, con);
-    );
-}
-
-void add_container_to_stack(struct container *con)
+struct monitor *container_get_monitor(struct container *con)
 {
     if (!con)
-        return;
+        return NULL;
+    if (con->client->m) {
+        return con->client->m;
+    }
 
+    struct workspace *ws = get_workspace(con->client->ws_id);
+    struct monitor *m  = ws->m;
+
+    if (ws->tagset) {
+        m = ws->tagset->m;
+    }
+    return m;
+}
+
+inline int absolute_x_to_container_relative(struct wlr_box geom, int x)
+{
+    return x - geom.x;
+}
+
+inline int absolute_y_to_container_relative(struct wlr_box geom, int y)
+{
+    return y - geom.y;
+}
+
+int get_position_in_container_stack(struct container *con)
+{
+    if (!con)
+        return INVALID_POSITION;
+
+    struct monitor *m = container_get_monitor(con);
+    struct tagset *tagset = monitor_get_active_tagset(m);
+    int position = find_in_composed_list(tagset->list_set->container_lists, cmp_ptr, con);
+    return position;
+}
+
+struct container *get_container_from_container_stack_position(int i)
+{
     struct monitor *m = selected_monitor;
-    if (con->client->type == LAYER_SHELL) {
-        switch (con->client->surface.layer->current.layer) {
-            case ZWLR_LAYER_SHELL_V1_LAYER_BACKGROUND:
-                g_ptr_array_insert(m->layer_visual_stack_background, 0, con);
-                break;
-            case ZWLR_LAYER_SHELL_V1_LAYER_BOTTOM:
-                g_ptr_array_insert(m->layer_visual_stack_bottom, 0, con);
-                break;
-            case ZWLR_LAYER_SHELL_V1_LAYER_TOP:
-                g_ptr_array_insert(m->layer_visual_stack_top, 0, con);
-                break;
-            case ZWLR_LAYER_SHELL_V1_LAYER_OVERLAY:
-                g_ptr_array_insert(m->layer_visual_stack_overlay, 0, con);
-                break;
-        }
-        return;
-    }
+    struct tagset *tagset = monitor_get_active_tagset(m);
+    struct container *con = get_container(tagset, i);
+    return con;
+}
 
-    if (con->floating) {
-        g_ptr_array_insert(m->floating_visual_stack, 0, con);
-        return;
-    }
+bool is_resize_not_in_limit(struct wlr_fbox *geom, struct resize_constraints *resize_constraints)
+{
+    bool is_width_not_in_limit = geom->width < resize_constraints->min_width ||
+        geom->height < resize_constraints->min_height;
 
-    g_ptr_array_insert(m->tiled_visual_stack, 0, con);
+    bool is_height_not_in_limit = geom->width > resize_constraints->max_width ||
+        geom->height > resize_constraints->max_height;
+
+
+    return is_width_not_in_limit || is_height_not_in_limit;
 }
 
 void set_container_workspace(struct container *con, struct workspace *ws)
@@ -750,84 +720,6 @@ void move_container_to_workspace(struct container *con, struct workspace *ws)
     ipc_event_workspace();
 }
 
-struct monitor *container_get_monitor(struct container *con)
-{
-    if (!con)
-        return NULL;
-    if (con->client->m) {
-        return con->client->m;
-    }
-
-    struct workspace *ws = get_workspace(con->client->ws_id);
-    struct monitor *m  = ws->m;
-
-    if (ws->tagset) {
-        m = ws->tagset->m;
-    }
-    return m;
-}
-
-void workspace_remove_container(struct workspace *ws, struct container *con)
-{
-    DO_ACTION(
-            remove_in_composed_list(list_set->container_lists, cmp_ptr, con);
-            );
-}
-
-void workspace_remove_container_from_focus_stack(struct workspace *ws, struct container *con)
-{
-    DO_ACTION(
-            remove_in_composed_list(list_set->focus_stack_lists, cmp_ptr, con);
-            );
-}
-
-void workspace_remove_independent_container(struct workspace *ws, struct container *con)
-{
-    DO_ACTION(
-            g_ptr_array_remove(list_set->independent_containers, con);
-            );
-}
-
-inline int absolute_x_to_container_relative(struct wlr_box geom, int x)
-{
-    return x - geom.x;
-}
-
-inline int absolute_y_to_container_relative(struct wlr_box geom, int y)
-{
-    return y - geom.y;
-}
-
-int get_position_in_container_stack(struct container *con)
-{
-    if (!con)
-        return INVALID_POSITION;
-
-    struct monitor *m = container_get_monitor(con);
-    struct tagset *tagset = monitor_get_active_tagset(m);
-    int position = find_in_composed_list(tagset->list_set->container_lists, cmp_ptr, con);
-    return position;
-}
-
-struct container *get_container_from_container_stack_position(int i)
-{
-    struct monitor *m = selected_monitor;
-    struct tagset *tagset = monitor_get_active_tagset(m);
-    struct container *con = get_container(tagset, i);
-    return con;
-}
-
-bool is_resize_not_in_limit(struct wlr_fbox *geom, struct resize_constraints *resize_constraints)
-{
-    bool is_width_not_in_limit = geom->width < resize_constraints->min_width ||
-        geom->height < resize_constraints->min_height;
-
-    bool is_height_not_in_limit = geom->width > resize_constraints->max_width ||
-        geom->height > resize_constraints->max_height;
-
-
-    return is_width_not_in_limit || is_height_not_in_limit;
-}
 
 bool container_is_bar(struct container *con)
 {
