+++ src/keybinding.c
@@ -4,6 +4,7 @@
 #include "tile/tileUtils.h"
 #include "utils/parseConfigUtils.h"
 #include "stringop.h"
+#include "workspace.h"
 
 const char *mods[8] = {"Shift_L", "Caps_Lock", "Control_L", "Alt_L", "", "", "Super_L", "ISO_Level3_Shift"};
 const char *modkeys[4] = {"Alt_L", "Num_Lock", "ISO_Level3_Shift", "Super_L"};
